+++ restli-tools/src/main/java/com/linkedin/restli/tools/idlcheck/CompatibilityLevel.java
@@ -28,6 +28,7 @@ public enum CompatibilityLevel
 {
   OFF,
   IGNORE,
+  EXTENSION,
   BACKWARDS,
   EQUIVALENT;